requirements.txt
numpy
pandas
seaborn
joblib
sklearn
pyyaml
openpyxl